gitlab_hook: renaming module name
